services/chat/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv

langchain
langchain-core
langchain-community
langchain-ollama
langchainhub

unstructured
pypdf
ollama
